requirements.txt
pandas
beautifulsoup4
openai
python-dotenv 